% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDataset}
\alias{myDataset}
\title{Show an alert}
\usage{
myDataset(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A valid shiny \code{session}.}

\item{msg}{Message to display.}
}
\description{
Show a vanilla JavaScript alert.
}
\examples{
library(shiny)

ui <- fluidPage(
  use_name(),
  verbatimTextOutput("response")
)

server <- function(input, output){
myDataset("Please enter something:")
 output$response <- renderPrint({
   input$myDatasetResponse
 })
}

if(interactive())
 shinyApp(ui, server)

}
