% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDatasetAddRow}
\alias{myDatasetAddRow}
\title{Add a row to the myDataset widget}
\usage{
myDatasetAddRow(
  id,
  row_id,
  name,
  size,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The id of the widget}

\item{row_id}{The id of the datasetID}

\item{name}{The name of the dataset}

\item{size}{The size of the dataset}

\item{session}{A valid shiny \code{session}.}
}
\description{
Add a row to the myDataset widget
}
