% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDatasetDownloadHandler}
\alias{myDatasetDownloadHandler}
\title{Download a dataset from the myDataset widget}
\usage{
myDatasetDownloadHandler(
  filename,
  extension,
  content,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{filename}{The output filename}

\item{extension}{The output file extension ("csv" or "json")}

\item{content}{The content of the file. This can be a string or a list.}

\item{session}{A valid shiny \code{session}.}
}
\description{
Download a dataset from the myDataset widget
}
