% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDatasetGetCookie}
\alias{myDatasetGetCookie}
\title{Get a cookie}
\usage{
myDatasetGetCookie(name, session = getDefaultReactiveDomain())
}
\arguments{
\item{name}{The name of the cookie}
}
\value{
cookie_value The value of the cookie
}
\description{
Get a cookie
}
