% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDatasetSetCookie}
\alias{myDatasetSetCookie}
\title{Set a cookie}
\usage{
myDatasetSetCookie(name, msg, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{name}{The name of the cookie}

\item{value}{The value of the cookie}
}
\description{
Set a cookie
}
