% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDatasetToggleDownload}
\alias{myDatasetToggleDownload}
\title{Blocks the download button until the download is complete}
\usage{
myDatasetToggleDownload(
  parentID,
  datasetID,
  action = "hide",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{parentID}{The id of the myDataset widget}

\item{datasetID}{The id of the dataset}

\item{action}{Can be 'hide' or 'show'. This will block or unblock the download button}
}
\description{
Blocks the download button until the download is complete
}
