% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDataset.R
\name{myDataset}
\alias{myDataset}
\title{Construct a myDataset widget}
\usage{
myDataset(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The id of the widget}

\item{session}{A valid shiny \code{session}.}
}
\description{
Construct a myDataset widget
}
